Refactor: Centralize `maxSteps` and `targetSteps` logic in UI state
This commit refactors the steps-related ViewModels and UI states to centralize the calculation and management of `maxSteps` and `targetSteps`. Instead of being collected as separate state flows in the views, these values are now calculated within the ViewModel and included directly in the `StepsDailyUiState` and `StepsOverviewUiState`.

This change simplifies the composable views by removing the need to collect multiple state flows and makes the UI state the single source of truth for the chart's configuration.

### Key Changes:

-   **`StepsDailyViewModel.kt` & `StepsOverviewViewModel.kt`:**
    -   A new private `maxSteps` state flow has been introduced, which combines `stepsDataFlow` and `targetSteps` to calculate the maximum value for the chart's Y-axis. It now ensures the chart accommodates both the highest recorded steps and the user's target.
    -   The `targetSteps` flow is now private and nullable (`Int?`), providing a `null` value if no profile is set.
    -   The main `uiState` flow now combines `stepsDataFlow`, `maxSteps`, and `targetSteps`, packaging them into `StepsDailyUiState` and `StepsOverviewUiState`.
    -   The public `maxSteps` state flow, which was previously derived from the `uiState`, has been removed.

-   **`StepsDailyUiState.kt` & `StepsOverviewUiState.kt`:**
    -   Added `maxSteps: Int` and `target: Int?` properties to hold the calculated maximum steps and target steps, respectively.

-   **`StepsDailyView.kt` & `StepsOverviewView.kt`:**
    -   Removed the collection of `maxSteps` and `targetSteps` as separate state flows.
    -   The `PlotSteps` and `StepsDetailView` composables now receive `maxSteps` and `targetSteps` directly from the `uiState` object.

-   **`PlotSteps.kt` & `PlotDailySteps.kt`:**
    -   Updated the `maxSteps` parameter type from `UInt` to `Int` to align with the new state and view model logic. The Y-axis model (`IntLinearAxisModel`) is now constructed with a range of `0..maxSteps`.
    -   The `targetSteps` parameter is now a nullable `Int?`.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/steps/PlotDailySteps.kt
@@ -52,7 +52,7 @@ fun PlotDailySteps(
     modifier: Modifier = Modifier,
     stepsData: List<Steps>,
     times: List<LocalTime>,
-    maxSteps: UInt,
+    maxSteps: Int,
     thumbnail: Boolean = false,
     targetSteps: Int? = null
 ) {
@@ -68,7 +68,7 @@ fun PlotDailySteps(
 
         XYGraph(
             xAxisModel = CategoryAxisModel(times),
-            yAxisModel = IntLinearAxisModel(0..maxSteps.toInt()),
+            yAxisModel = IntLinearAxisModel(0..maxSteps),
             horizontalMajorGridLineStyle = null,
             horizontalMinorGridLineStyle = null,
             verticalMajorGridLineStyle = null,
